# Conflicts:
#	Project.sdf
#	Project.suo
#	Project/Account1.h
#	Project/MYSQL.cpp
#	Project/MYSQL.h
#	Project/Project.vcxproj.filters
#	Project/Release/Account1.obj
#	Project/Release/CL.read.1.tlog
#	Project/Release/CL.write.1.tlog
#	Project/Release/MYSQL.obj
#	Project/Release/MainWindow.obj
#	Project/Release/PosUI.obj
#	Project/Release/Project.log
#	Project/Release/Project.obj
#	Project/Release/Project.vcxproj.GenerateResource.Cache
#	Project/Release/SaleSystemForm.obj
#	Project/Release/Staff1.obj
#	Project/Release/WareHouseForm.obj
#	Project/Release/WarehouseDashboard.obj
#	Project/Release/cl.command.1.tlog
#	Project/Release/link.command.1.tlog
#	Project/Release/link.read.1.tlog
#	Project/Release/link.write.1.tlog
#	Project/Release/vc100.pdb
#	Release/Project.exe
#	Release/Project.pdb

+++ Project/PosUI.h
@@ -258,7 +258,7 @@ namespace Project {
 			productName = gcnew array<String^> (productList.size());
 			priceForFew = gcnew array<Double> (productList.size());
 			array_size = productList.size();
-			for(int i=0;i<productList.size();i++){
+			for(int i=0;i<(int)productList.size();i++){
 				String ^ items = gcnew String(productList[i].getName().c_str());
 				productName[i] = items;
 				productID[i] = productList[i].getID();
